bug fixed, source found
if a ball is directly on another ball, it is incabable of moving out of the way of the way along its normal, because when impulse is caused, only its x velocity changes. After a while, the above ball's y velocity builds up due to gravity, until it gets to a point where the 'collision' is occurring directly on top of the under ball. When the collision normal is found, a 0 vector is returned, causing a divide by zero error (which has now been fixed)